MAYA Makerbot Agent-Adaper initial commit
Committing the version of the code built by MAYA Design, tested on a
Makerbot Replicator 2. This code was built for MTConnect version 1.2.

This version is a functioning prototype, meaning all functionality
should work though their may be pieces that are missing due to the
nature of it being a built solely for the purpose of discovering the
process of building an MTConnect Agent.